split regex module
